cleaning up unnecessary vue files
