+++ packages/frontend-2/pages/projects/[id]/index.vue
@@ -42,10 +42,14 @@ graphql(`
   fragment ProjectPageProject on Project {
     id
     createdAt
+    modelCount: models(limit: 0) {
+      totalCount
+    }
+    commentThreadCount: commentThreads(limit: 0) {
+      totalCount
+    }
     ...ProjectPageProjectHeader
     ...ProjectPageTeamDialog
-    ...ProjectPageLatestItemsModels
-    ...ProjectPageLatestItemsComments
   }
 `)
